feat(edit.ts): recursive search element enable
