+++ src/utils/main.tsx
@@ -1,0 +1,5 @@
+export const createZestyDiv = () => {
+   const elem = document.createElement("div")
+   elem.id = "zesty-explorer"
+   document.body.appendChild(elem)
+}
